requirements.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.4
Django==3.1.2
django-allauth==0.44.0
django-cors-headers==3.7.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
idna==2.10
install==1.3.4
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.1
Pillow==8.2.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
